Add ChatMarkdownContentPart
