+++ src/vs/workbench/contrib/chat/browser/chatContentParts/chatProgressContentPart.ts
@@ -3,12 +3,13 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+import { alert } from 'vs/base/browser/ui/aria/aria';
 import { Codicon } from 'vs/base/common/codicons';
 import { MarkdownString } from 'vs/base/common/htmlContent';
 import { Disposable } from 'vs/base/common/lifecycle';
 import { ThemeIcon } from 'vs/base/common/themables';
 import { MarkdownRenderer } from 'vs/editor/browser/widget/markdownRenderer/browser/markdownRenderer';
-import { IChatCommandButton, IChatProgressMessage, IChatTask } from 'vs/workbench/contrib/chat/common/chatService';
+import { IChatProgressMessage, IChatTask } from 'vs/workbench/contrib/chat/common/chatService';
 
 export class ChatProgressContentPart extends Disposable {
 	public readonly element: HTMLElement;
@@ -33,14 +34,5 @@ export class ChatProgressContentPart extends Disposable {
 		result.element.classList.add('progress-step');
 
 		this.element = result.element;
-	}
-
-	isDifferent(command: IChatCommandButton): boolean {
-		// This chat type can't be updated in the model
-		return false;
-	}
-
-	update(command: IChatCommandButton): void {
-		// This chat type can't be updated in the model
 	}
 }
